Actions: Add an addaction drawing an icon on vehicle's spawn point location.
Currently players need to remember exactly which spawn point they got a vehicle from if they want to swap it for another (need to swap from the exact same spawn point it was created at).

This can cause issues when there are empty spawn points.

1. Player RTBs with the vehicle, sees empty spawn point.
2. Player lands/parks their vehicle on the empty spawn point.
3. Vehicle that was spawned at that spawn point then respawns, causing an explosion.

This gives players a simple utility to check where their vehicle's spawn point is located so they can return it properly.

Or they can blow it up / wreck and recover it. that works too.