Khe Sanh Update
Updated Khe Sanh to new standard.

Overhaul; way too many updates to list.

Overview:
Spike Teams to the north of the runway.
Air Force on the west end of the runway.
ACAV is west of Mike Force, and is surrounded by Marines to the south, Blackhorse to the east, and ARVN to the west.
7th, Muskets, MACV, and MPs is East of Mike Force.
Mike Force is now the place to go to load onto helos.
Green Hornets is where you spawn to get a helo.

Added Dock teleporter to options, UDT is up at Khe Sanh, with the dock down at the river.

Added Teleporter and destination to CamLaoNam, for continuity.

KNOWN ISSUE: Carrier Spawn from Teleporter is ... not working as intended.